FEED Edit, Delete DONE//2024.11.19

+++ src/main/java/app/labs/linksy/Repository/feed_create_repository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface feed_create_repository extends CrudRepository<Feed, Long> {
+public interface feed_create_repository extends CrudRepository<Feed, Integer> {
     // 기본적인 CRUD 메서드를 사용하기 위해 CrudRepository 상속
 }
